Created email validation checker.
Implemented a routine isValid() to check if the email address has a valid format. In this routine I am using The RE library (Regular Expression) which allows the user to match character and see if all characters in email align with the RE variable format. If true return true otherwise, false.